feat: Widget Framework + dep fixes (x86_64 0.14.12 + uart_16550)
Widget Framework (widget.rs) — Taylor, 186 lines — Label, Button, TextInput with events
Fixed: reverted to x86_64 0.14.12 for instructions::port support
Fixed: re-added uart_16550 for serial.rs compatibility
Fixed: GDT add_entry, IDT indexing

13,500+ lines, all compiling

+++ kernel/src/interrupts.rs
@@ -61,7 +61,7 @@ static IDT: spin::Lazy<InterruptDescriptorTable> = spin::Lazy::new(|| {
             .set_stack_index(crate::gdt::DOUBLE_FAULT_IST_INDEX);
     }
 
-    idt[InterruptIndex::Keyboard as u8].set_handler_fn(keyboard_interrupt_handler);
+    idt[InterruptIndex::Keyboard as u8 as usize].set_handler_fn(keyboard_interrupt_handler);
 
     idt
 });
